Remove extra layer
